commit correlation Id

+++ src/main/java/Main.java
@@ -21,7 +21,7 @@ public class Main {
         } finally {
             try {
                 if (clientSocket != null) {
-                    int messageSize = 0x00000000; //4 bytes
+                    long messageSize = 0x00000000; //4 bytes
                     long correlationId = 0x00000007;
                     String response = String.valueOf(messageSize + correlationId);
                     clientSocket.getOutputStream().write(response.getBytes());
